langchain-template/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-anthropic
langchain-core
xpander-sdk
openai
anthropic
